+++ src/lib/UI/domainComponents/cards/RecipeCard.svelte
@@ -74,7 +74,7 @@
   }
 
   h5 {
-    @apply text-2xl font-bold tracking-tight text-gray-900 dark:text-white;
+    @apply text-2xl font-bold tracking-tight text-gray-900;
   }
 
   .label {
@@ -82,7 +82,7 @@
   }
 
   .content {
-    @apply inline font-normal dark:text-gray-400;
+    @apply inline font-normal;
     white-space: pre-line;
   }
 </style>