+++ src/lib/UI/domainComponents/forms/CoffeeBeansSelect.svelte
@@ -86,7 +86,7 @@
       {/if}
     </div>
   </div>
-  <p class="text-sm text-red-600 dark:text-red-500" class:mt-2={validationMessage.length > 0}>{validationMessage}</p>
+  <p class="text-sm text-red-600" class:mt-2={validationMessage.length > 0}>{validationMessage}</p>
 </div>
 
 <style lang="postcss">
@@ -110,13 +110,11 @@
   .valid {
     @apply bg-gray-50 border border-gray-300 text-gray-900 text-base rounded-lg p-2.5;
     @apply focus:ring-blue-500 focus:border-blue-500;
-    @apply dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white;
-    @apply dark:focus:ring-blue-500 dark:focus:border-blue-500;
   }
 
   .invalid {
     @apply bg-red-50 border border-red-500 text-red-900 placeholder-red-700 text-base rounded-lg focus:ring-red-500;
-    @apply focus:border-red-500 block w-full p-2.5 dark:bg-red-100 dark:border-red-400;
+    @apply focus:border-red-500 block w-full p-2.5;
     background-color: #fef2f2;
   }