+++ src/lib/UI/domainComponents/modals/EditCoffeeBeansModal.svelte
@@ -5,7 +5,7 @@
 <script lang="ts">
   import { tick } from "svelte";
   import { goto } from "$app/navigation";
-  import { editCoffeeBeans } from "$lib/database/manageCoffeeBeans";
+  import { checkCoffeeBeansDuplicate, editCoffeeBeans } from "$lib/database/manageCoffeeBeans";
   import { routes } from "$lib/domain/constants/routes";
   import { CoffeeBeans, CoffeeBeansEditSubmit } from "$lib/domain/entities/CoffeeBeans";
   import Label from "$lib/UI/genericComponents/forms/Label.svelte";
@@ -49,7 +49,7 @@
   let hasUnsavedChanges: boolean = false;
 
   // Unsaved changes reactivity:
-  $: if (item && name === item.name && description === item.description) {
+  $: if (item && name.trim() === item.name && description.trim() === item.description) {
     hasUnsavedChanges = false;
   } else {
     hasUnsavedChanges = true;
@@ -72,6 +72,9 @@
   }
 
   async function handleSubmit() {
+    // Trim the name and description in the form too:
+    name = name.trim();
+    description = description.trim();
     // Validate and save the edited coffee beans:
     const editedCoffeeBeans: CoffeeBeans = new CoffeeBeans({ id: item.id, name, description });
     const coffeeBeansSubmit: CoffeeBeansEditSubmit | "ValidationFailed_NameMustBeAtLeast3CharsLong" =
@@ -88,7 +91,7 @@
     // Guard clauses:
     if (coffeeBeans === "Failure_NameAlreadyExist") {
       nameValidationFailed = true;
-      validationMessage = "Coffee beans with this name exist already.";
+      validationMessage = "Coffee beans with this name already exist.";
       return;
     }
     if (coffeeBeans === "CoffeeBeansNotFound") {
@@ -119,10 +122,30 @@
   }
 
   function handleInputChange() {
-    if (nameValidationFailed) {
-      nameValidationFailed = false;
-      validationMessage = "";
+    if (CoffeeBeans.hasValidName({ name }) === "ValidationFailed_NameMustBeAtLeast3CharsLong") {
+      nameValidationFailed = true;
+      validationMessage = "Name must be at least 3 characters long.";
+      return;
     }
+    if (name.trim().toLowerCase() === item.name.toLowerCase()) {
+      nameValidationFailed === false;
+      validationMessage = "";
+      return;
+    }
+    checkCoffeeBeansDuplicate(name).then((value) => {
+      switch (value) {
+        case "CoffeeBeansNotFound":
+          nameValidationFailed = false;
+          validationMessage = "";
+          break;
+        case "Failure_NameAlreadyExist":
+          nameValidationFailed = true;
+          validationMessage = `Coffee beans "${name.trim()}" already exist.`;
+          break;
+        default:
+          break;
+      }
+    });
   }
 
   function handleSaveButtonTabKeydown(event: KeyboardEvent) {
@@ -157,12 +180,12 @@
         type="text"
         bind:this={inputDom}
         bind:value={name}
-        class:unsaved-changes={item.name !== undefined && item.name !== name}
+        class:unsaved-changes={item.name !== undefined && name.trim() !== item.name}
         on:focusin={handleInputFocusIn}
         on:input={handleInputChange}
         on:keydown={handleEnterKey}
       />
-      <p class="mt-2 text-sm text-red-600 dark:text-red-500">{validationMessage}</p>
+      <p class="mt-2 text-sm text-red-600">{validationMessage}</p>
     </div>
     <div class="my-div mb-5">
       <Textarea
@@ -179,7 +202,7 @@
     </div>
     <button
       class="button-submit"
-      disabled={hasUnsavedChanges === false || CoffeeBeans.hasValidName({ name }) !== true}
+      disabled={nameValidationFailed || hasUnsavedChanges === false || CoffeeBeans.hasValidName({ name }) !== true}
       type="submit"
       bind:this={saveButtonDOM}
       on:keydown={handleSaveButtonTabKeydown}
@@ -193,13 +216,11 @@
   .input-name {
     @apply bg-gray-50 border border-gray-300 text-gray-900 text-base rounded-lg block w-full p-2.5;
     @apply focus:ring-blue-500 focus:border-blue-500;
-    @apply dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white;
-    @apply dark:focus:ring-blue-500 dark:focus:border-blue-500;
   }
 
   .input-name-validation-failed {
     @apply border border-red-500 text-red-900 placeholder-red-700 text-base rounded-lg focus:ring-red-500;
-    @apply focus:border-red-500 block w-full p-2.5 dark:bg-red-100 dark:border-red-400;
+    @apply focus:border-red-500 block w-full p-2.5;
     background-color: #fef2f2;
   }
 
@@ -210,8 +231,7 @@
 
   .button-submit {
     @apply text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium;
-    @apply rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700;
-    @apply dark:focus:ring-blue-800;
+    @apply rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center;
     width: 100%;
     margin-left: 0;
     margin-right: 0;