+++ src/lib/UI/domainComponents/pageBlocks/AddDemoCoffeeBeans_PageBlock.svelte
@@ -52,8 +52,6 @@
   .add-demo-coffee-beans {
     @apply text-yellow-400 hover:text-white border border-yellow-400 hover:bg-yellow-500 focus:ring-4;
     @apply focus:outline-none focus:ring-yellow-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center;
-    @apply dark:border-yellow-300 dark:text-yellow-300 dark:hover:text-white dark:hover:bg-yellow-400;
-    @apply dark:focus:ring-yellow-900;
     margin-top: 1rem;
   }
 </style>