+++ src/lib/UI/genericComponents/Card.svelte
@@ -9,8 +9,7 @@
 
 <style lang="postcss">
   a {
-    @apply block bg-white border border-gray-200 rounded-lg shadow dark:bg-gray-800;
-    @apply dark:border-gray-700 dark:hover:bg-gray-700;
+    @apply block bg-white border border-gray-200 rounded-lg shadow;
     margin-top: 12px;
     margin-bottom: 12px;
     padding: 1rem 1.5rem 1.5rem 1.5rem;