+++ src/lib/UI/layout/components/HamburgerMenu.svelte
@@ -70,8 +70,7 @@
 
   .menu-button {
     @apply inline-flex items-center p-2 w-10 h-10 justify-center text-sm text-gray-500 rounded-lg md:hidden;
-    @apply hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400;
-    @apply dark:hover:bg-gray-700 dark:focus:ring-gray-600;
+    @apply hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200;
   }
 
   .menu-div {
@@ -86,14 +85,11 @@
 
   .ul-menu {
     @apply flex flex-col p-4 md:p-0 mt-4 font-medium border border-gray-100 rounded-lg bg-gray-50 md:space-x-8;
-    @apply rtl:space-x-reverse md:flex-row md:mt-0 md:border-0 md:bg-white dark:bg-gray-800 md:dark:bg-gray-900;
-    @apply dark:border-gray-700;
+    @apply rtl:space-x-reverse md:flex-row md:mt-0 md:border-0 md:bg-white;
   }
 
   .a-menu {
     @apply block py-2 px-3 text-white bg-blue-700 rounded md:bg-transparent md:text-blue-700 md:p-0;
-    @apply md:dark:text-blue-500;
-
     margin-top: 8px;
     margin-bottom: 8px;
   }