+++ src/lib/prevVersions/v3/core/CoffeeBeansV3.ts
@@ -1,0 +1,5 @@
+export interface ICoffeeBeans_v3 {
+  id: number;
+  name: string;
+  description: string;
+}