+++ src/lib/prevVersions/v3/database/CoffeeBeansDBv3.ts
@@ -1,0 +1,7 @@
+export interface ICoffeeBeansDB_v3 {
+  id: number;
+  name: string;
+  nameLowerCase: string;
+  description: string;
+  softDeletionTimestamp: number | undefined;
+}