+++ src/lib/prevVersions/v4/database/EnhancedCoffeeBeansDBv4.ts
@@ -1,0 +1,6 @@
+export interface IEnhancedCoffeeBeansDB_v4 {
+  id: number;
+  recipesCount: number;
+  latestRecipeTimestamp: number | undefined;
+  earliestRecipeTimestamp: number | undefined;
+}