+++ src/lib/prevVersions/v4/database/RecipeDBv4.ts
@@ -1,0 +1,12 @@
+export interface IRecipeDB_v4 {
+  id: number;
+  coffeeBeansId: number;
+  recipeTarget: string;
+  recipeResult: string;
+  recipeThoughts: string;
+  favorite: boolean;
+  rating: number;
+  outWeight: number;
+  timestamp: number;
+  softDeletionTimestamp: number | undefined;
+}